i915: Add support for reprobing for a panel
In some situations we may boot in a situation where a panel is attached to
a mux, the mux is not routed to i915 and the driver for switching the mux
only binds after i915 is loaded. In that case we need to reprobe for an
attached panel. Add support to do so.

Signed-off-by: Matthew Garrett <matthew.garrett@nebula.com>