vga_switcheroo: Allow stashing of panel data
Not all MUXes allow us to connect the panel data channel to a GPU without
handing over the entire panel and triggering additional flickering during
boot. We only need to do this in order to probe for data that the first
GPU driver has already identified, so add some functions for stashing that
data in vga_switcheroo where it can be retrieved by the other driver later.

Signed-off-by: Matthew Garrett <matthew.garrett@nebula.com>